backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
Pillow
transformers
datasets
accelerate
torch
safetensors
genai
httpx
